requirements.txt
Django>=4.2,<5
djangorestframework
django-filter
djangorestframework-simplejwt
python-dotenv
django-cors-headers
